Dev task for npm, update demo to json-server
