+++ package.json
@@ -22,6 +22,7 @@
     "date-fns": "^2.30.0",
     "dayjs": "^1.11.10",
     "express": "^4.18.2",
+    "finalhandler": "^1.2.0",
     "formik": "^2.2.9",
     "framer-motion": "^4.1.17",
     "js-yaml": "^4.1.0",
@@ -46,6 +47,7 @@
     "react-syntax-highlighter": "^15.5.0",
     "redux": "^4.2.0",
     "semver-compare": "^1.0.0",
+    "serve-static": "^1.15.0",
     "simplebar": "^5.3.8",
     "simplebar-react": "^2.4.1",
     "timeago.js": "^4.0.2",
@@ -81,7 +83,9 @@
   ],
   "scripts": {
     "client": "vite",
-    "client-build": "vite build --base=/cosmos-ui/",
+    "webpack:setup-env": "cross-env withReport=$npm_config_with_report analyzeDeps=$npm_config_analyze_deps",
+    "webpack:build": "npm run webpack:setup-env -- webpack --config webpack.prod.js --progress",
+    "webpack:start": "npm run webpack:setup-env -- webpack serve --open --config webpack.dev.js",
     "start": "env COSMOS_CONFIG_FOLDER=/mnt/e/work/Cosmos-Server/zz_test_config/ CONFIG_FILE=./config_dev.json EZ=UTC ACME_STAGING=true build/cosmos",
     "build": "sh build.sh",
     "dev": "npm run build && npm run start",
@@ -116,7 +120,9 @@
     "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/react": "^14.1.2",
     "@types/js-yaml": "^4.0.9",
+    "@types/json-server": "^0.14.7",
     "@types/lodash.merge": "^4.6.9",
+    "@types/node": "^20.10.2",
     "@types/react-color": "^3.0.10",
     "@types/webpack-bundle-analyzer": "^4.6.3",
     "babel-loader": "^9.1.3",
@@ -124,10 +130,13 @@
     "babel-plugin-import": "^1.13.8",
     "babel-plugin-transform-runtime": "^6.23.0",
     "buffer": "^6.0.3",
+    "cross-env": "^7.0.3",
+    "cross-var": "^1.1.0",
     "crypto-browserify": "^3.12.0",
     "css-loader": "^6.8.1",
     "css-minimizer-webpack-plugin": "^5.0.1",
     "file-loader": "^6.2.0",
+    "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.3",
     "image-minimizer-webpack-plugin": "^3.8.3",
     "imagemin": "^8.0.1",
@@ -136,6 +145,7 @@
     "imagemin-optipng": "^8.0.0",
     "imagemin-svgo": "^10.0.1",
     "inspectpack": "^4.7.1",
+    "json-server": "^0.17.4",
     "mini-css-extract-plugin": "^2.7.6",
     "path-browserify": "^1.0.1",
     "prettier": "2.7.1",
@@ -149,6 +159,7 @@
     "webpack": "^5.89.0",
     "webpack-bundle-analyzer": "^4.10.1",
     "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
     "webpack-merge": "^5.10.0"
   }
 }
